refactor: centralize phrase or joke
