+++ game_board15/handlers.py
@@ -20,9 +20,8 @@ from logic.phrases import (
     SELF_HIT,
     SELF_KILL,
     SELF_MISS,
-    random_phrase,
-    random_joke,
 )
+from .utils import _phrase_or_joke
 import random
 import asyncio
 
@@ -53,17 +52,6 @@ def _keyboard() -> InlineKeyboardMarkup:
         ]
     ]
     return InlineKeyboardMarkup(arrows + matrix + actions)
-
-
-def _phrase_or_joke(match, player_key: str, phrases: list[str]) -> str:
-    shots = match.shots[player_key]
-    start = shots.get("joke_start")
-    if start is None:
-        start = shots["joke_start"] = random.randint(1, 10)
-    count = shots.get("move_count", 0)
-    if count >= start and (count - start) % 10 == 0:
-        return f"Слушай анекдот:\n{random_joke()}\n\n"
-    return f"{random_phrase(phrases)} "
 
 
 async def board15(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
