+++ game_board15/router.py
@@ -15,20 +15,8 @@ from logic.phrases import (
     SELF_HIT,
     SELF_KILL,
     SELF_MISS,
-    random_phrase,
-    random_joke,
 )
-
-
-def _phrase_or_joke(match, player_key: str, phrases: list[str]) -> str:
-    shots = match.shots[player_key]
-    start = shots.get("joke_start")
-    if start is None:
-        start = shots["joke_start"] = random.randint(1, 10)
-    count = shots.get("move_count", 0)
-    if count >= start and (count - start) % 10 == 0:
-        return f"Слушай анекдот:\n{random_joke()}\n\n"
-    return f"{random_phrase(phrases)} "
+from .utils import _phrase_or_joke
 
 
 async def _send_state(context: ContextTypes.DEFAULT_TYPE, match, player_key: str, message: str) -> None:
