+++ game_board15/utils.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import random
+from logic.phrases import random_phrase, random_joke
+
+
+def _phrase_or_joke(match, player_key: str, phrases: list[str]) -> str:
+    shots = match.shots[player_key]
+    start = shots.get("joke_start")
+    if start is None:
+        start = shots["joke_start"] = random.randint(1, 10)
+    count = shots.get("move_count", 0)
+    if count >= start and (count - start) % 10 == 0:
+        return f"Слушай анекдот:\n{random_joke()}\n\n"
+    return f"{random_phrase(phrases)} "
+
+
+__all__ = ["_phrase_or_joke"]
